Restructured the src/ folder
The src folder is now structured into lib/ and angular/ folders, paving
the way for extracting functionality into non-angular modules.
