+++ src/angular/endev/directives/click.js
