+++ src/angular/endev/directives/edit.js
