+++ src/angular/endev/directives/endevAnnotation.js
