+++ src/angular/endev/directives/endevItem.js
