+++ src/angular/endev/directives/insertInto.js
