+++ src/angular/endev/directives/list.js
