+++ src/angular/endev/directives/object.js
