+++ src/angular/endev/directives/removeFrom.js
