+++ src/angular/endev/directives/value.js
