+++ src/angular/endev/templates/annotations.html
