+++ src/index.js
@@ -3,10 +3,10 @@ window._        = require("underscore");
 window.firebase = require("firebase");
 require("angularfire");
 
-require("./utils");
-require("./xml2json");
-require("./modules/datatag/data-tag");
-require("./modules/endev/endev");
+require("./lib/utils");
+require("./lib/xml2json");
+require("./angular/datatag/");
+require("./angular/endev/");
 
 angular.element(document).ready(function() {
     angular.bootstrap(document, ['Endev']);
